+++ src/gdal_algs.cpp
@@ -1433,7 +1433,13 @@ SpatVector SpatRaster::polygonize(bool round, bool values, bool narm, bool aggre
 	}
 
     GDALDataset *poDS = NULL;
+
+#if (GDAL_VERSION_MAJOR == 3 && GDAL_VERSION_MINOR >= 11) || (GDAL_VERSION_MAJOR >= 4)	
+    GDALDriver *poDriver = GetGDALDriverManager()->GetDriverByName( "MEM" );
+#else
     GDALDriver *poDriver = GetGDALDriverManager()->GetDriverByName( "Memory" );
+#endif
+
     if( poDriver == NULL )  {
         out.setError( "cannot create output driver");
         return out;